ADAFRUIT HT16K33 Custom Character Generator
This code will help you to generate any custom character directly from the computer/machine. You have to design the character on the GUI and that code will be transferred to Arduino which will code the Display for that particular character. No need to hardcode different characters again and again.